+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - PR #116 - Fix stream usage on CuPy raw kernels
 - PR #124 - Remove cp_stream and autosync
 - PR #127 - Fix selected documentation formatting errors
+- PR #138 - Fix overflow issues in `upfirdn`
 
 
 # cuSignal 0.14 (03 Jun 2020)